Simplify UriParser
The reverts https://github.com/apache/commons-vfs/pull/396 and related
changes and implements the same in a simpler way by replacing the
encoded characters already in `fixSeparators`.

This approach has a slightly higher risk at breaking existing behaviour,
but a lower risk of remaining problems in this part of the codebase. All
testcases still succeed.

This PR is intended to replace #543 and #555

This reverts commit cb45c9453669eb2570f582f491cb0e06b0e63466.
This reverts commit 5399c766287a06ce002beaed89fa8af5461e15d0.